requirements.txt
click==8.1.8
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==2.2.4
opencv-python==4.11.0.86
pillow==11.2.1
